PHAN-III-XE-CUA-2/requirements.txt
google-generativeai>=0.3.0
tkinter-tooltip>=2.0.0